Add ExitWindowsEx()

+++ WindowsAPI/MFStartup.cpp
@@ -4,5 +4,5 @@
 #include "com_khopan_win32_Win32.h"
 
 JNIEXPORT jlong JNICALL Java_com_khopan_win32_Win32_MFStartup(JNIEnv* environment, jclass win32, jlong Version, jlong dwFlags) {
-	return MFStartup(Version, dwFlags);
+	return MFStartup(static_cast<ULONG>(Version), static_cast<DWORD>(dwFlags));
 }
